+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 ## 1.13.1 (2017-02-24)
 - FEATURE: version file added
+- FEATURE: now docker engine start automatically at boot
 
 ## 1.13.1 (2017-02-11)
 - UPDATE: update to docker 1.13.1 on the IaaS